+++ tests/broker/test_broker.py
@@ -300,3 +300,4 @@ async def test_broker_client(example_device, shvbroker):
 
 
 # TODO test clients disconnect on idle
+# TODO test access level